+++ examples/prereqs_quickstart/aws-vpc/variables.tf
@@ -1,25 +1,11 @@
-variable "aws_region" {
-  description = "AWS region to deploy resources into"
-  type        = string
-  default     = "us-east-1"
-}
-
 variable "azs" {
   description = "availability zones to use in AWS region"
   type        = list(string)
-  default = [
-    "us-east-1a",
-    "us-east-1b",
-    "us-east-1c",
-  ]
 }
 
 variable "common_tags" {
   type        = map(string)
   description = "Tags for VPC resources"
-  default = {
-    Vault = "dev"
-  }
 }
 
 variable "resource_name_prefix" {
@@ -37,14 +23,6 @@ variable "private_subnet_cidrs" {
   ]
 }
 
-variable "private_subnet_tags" {
-  type        = map(string)
-  description = "Tags for private subnets. Be sure to provide these tags to the Vault installation module."
-  default = {
-    Vault = "deploy"
-  }
-}
-
 variable "public_subnet_cidrs" {
   description = "CIDR blocks for public subnets"
   type        = list(string)
@@ -60,3 +38,4 @@ variable "vpc_cidr" {
   type        = string
   default     = "10.0.0.0/16"
 }
+