+++ examples/prereqs_quickstart/versions.tf
@@ -1,5 +1,5 @@
 terraform {
-  required_version = ">= 0.15"
+  required_version = ">= 1.2.1"
 
   required_providers {
     aws = ">= 3.0.0, < 4.0.0"